+++ modules/internal/src/main/kotlin/org/polyfrost/oneconfig/internal/ui/themes/Provider.kt
@@ -2,10 +2,12 @@ package org.polyfrost.oneconfig.internal.ui.themes
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.compositionLocalOf
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
+import androidx.compose.runtime.withFrameNanos
 import androidx.compose.ui.graphics.Color
 import org.polyfrost.oneconfig.internal.ThemeConfig
 
@@ -20,6 +22,14 @@ val LocalTheme = compositionLocalOf<UITheme> { error("A UI theme is required but
 @Composable
 fun Theme(content: @Composable () -> Unit) {
     _accent = Color(ThemeConfig.accentColor.argb)
+
+    LaunchedEffect(Unit) {
+        while (true) {
+            withFrameNanos { }
+            if (ThemeConfig.accentColor.chroma) updateAccent()
+        }
+    }
+
     val target = ThemeRegistry.activeTheme ?: error("No active theme provided")
     CompositionLocalProvider(
         LocalTheme provides animateTheme(target),