+++ CHANGELOG.md
@@ -1,3 +1,12 @@
-1.0.0-beta.4 changelogs:
+1.0.0-beta.(wip)5 changelogs:
 
-- Fix test HUD being included by accident
+- Round non-SVG icons automatically (#659)
+
+- Fix duplicate mod card when a mod has both OneConfig and Mod Menu config support (#663)
+- Fix UI clipping sidebar (#665)
+- Fix UI keybind working in other GUIs
+- Fix HUD resizing not being correctly proportional to mouse drag (#662)
+- Fix MultiSelectDropdown not saving (#664)
+- Fix chroma UI accent color not working (#642)
+- Fix subcategories not being displayed as accordions in Cloth Config (#639)
+- Fix YACL compat not using min/max values (#669)